segment-param-1
part 1 of making page size possibly variable:
make Segment PTEMAPMEM a variable, based on the size of each page,
represented by log2(page size). larger pages in a given segment will cause
each of its Ptes to map bigger regions as a whole.

a later part will move knowledge of the indexing calculation into segment.c,
and most use of PGSZ will be confined to k10, although it still remains as the
smallest possible [logical] page size for a given architecture.

some #defines and other definitions that become redundant
will be removed later on.

[The soff argument to pio changed from a ulong to uintptr.
Pio is local to fault.c now, could be removed from portfns.h]
[The newsize variable in ibrk changed from long to usize]

Author(s): forsyth, jmk

Reference: 018-segment-param-1
Date: Wed Mar 13 15:46:23 GMT 2013

+++ port/portfns.h
@@ -210,7 +210,7 @@ void		pexit(char*, int);
 void		pgrpcpy(Pgrp*, Pgrp*);
 void		pgrpnote(ulong, char*, long, int);
 int		psindex(int);
-void		pio(Segment*, uintptr, ulong, Page**);
+void		pio(Segment*, uintptr, uintptr, Page**);
 #define		poperror()		up->nerrlab--
 int		postnote(Proc*, int, char*, int);
 int		pprint(char*, ...);
